Split Header and NavigationMenu implementation
